+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - `Csub` (use `cowel_sub` instead)
   - `Cmul` (use `cowel_mul` instead)
   - `Cdiv` (use `cowel_div_to_zero` instead)
+  - `literally` (use `cowel_source_as_text` instead)
+  - `unprocessed` (use `cowel_no_invoke` instead)
 - change the syntax of directive arguments (#87, #107, #109, #111, #112, #121)
   - there are now values/literals of specific types
   - only specific strings are allowed to be unquoted, rather than this being the default